[Refactor] 코드 에디터 커스텀 훅 분리 (초안)
- useTextSelection, useHighlights 훅으로 분리하여 모듈화
